Minimum-Viable Connector demo
NOTE, this connector implementation is completely unidirectional:
The connector assumes that the WebUI is already loaded and waiting for a capabilities feed.
It will not respond to capabilities requests, or feed requests.
It is always providing the Acceleration feed, at an interval of 1 second.
Also note that maximum compiler optimization is necessary to avoid the 16k IAR Workbench kickstart limit.
Without more systemic and manual optimization, we don't have room for a floating point formatter.
Some slightly ugly integer math was used in it's place...